Increased the resolution of images and changed draw behaviour
Rescaled the images to 500x500 (which should work on all phones) and changed the games behaviour on a draw to the following:

When your card causes a draw, you should be the next person to select an attribute. If you have more than two players however and your cards selected attribute is lower than the other players cards AND the winning cards are in a draw, one of the players, that is tied in this draw should be allowed to select the next attribute. This player is determined randomly, but again, only if the previous current player is not elligible to participate in the draw round.