Started with animations
Now when the submit button is pressed all players cards are put to the back, you see the other players cards flip open, then close and then your card gets updated and  it moves to the front.
The start of the game now looks a little weird and needs revision and the information about the enemy cards is not being passed yet.