Information is being passed, however still buggy
For some reason the host sees a completely random card for his enemy cards, for the clients it seems to work however.
The flipAllCards method does not seem to work for more than 2 players and could shed more light on this issue if it were fixed.
The game currently crashes if you play after another player lost and an easy fix would be an "empty" card with the background logo as image with an id of 0, so this will be done as well.
The compared message should also be given the information who is part of the draw round, so that those players cards don't get flipped, which could also be achieved with an empty card.