Quick fix for presentation
I could not manage to signal the winner properly in time, so I included a quick fix that works with 2 players only. Since 3 players has some graphical issues anyway it should be ok.

In order to fix this we need to assign each players ID to the enemy fragment views, with the own ID being assigned to the player card view and proceed from there.